use fabs() for floating point Original: 1bc47a11080da9e72dbf9a1f015046e6fa0d45b5

+++ lalapps/src/pulsar/TwoSpect/fastchisqinv.c
@@ -253,7 +253,7 @@ REAL8 cdf_ugaussian_Pinv(REAL8 P)
    XLAL_CHECK_REAL8( P != 1.0, XLAL_EFPOVRFLW );
    XLAL_CHECK_REAL8( P != 0.0, XLAL_EFPOVRFLW );
 
-   if (fabsf(dP) <= 0.425) {
+   if (fabs(dP) <= 0.425) {
       x = twospect_small(dP);
       return x;
    }
